+++ frontend/src/app/(no-websocket)/layout.tsx
@@ -1,11 +1,11 @@
-"use client";
-// import type { Metadata } from "next";
+import ClientLayout from "@/app/(no-websocket)/layout-client";
+import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
-import "../globals.css";
-import { usePathname } from "next/navigation";
-import { useEffect } from "react";
-import { ProtectedRoute } from "./auth/hooks/ProtectedRoute";
-import { useAuthStore } from "./auth/store/useAuthStore";
+
+export const metadata: Metadata = {
+  title: "Quizoo",
+  description: "Real-time quiz app with a cute animals theme",
+};
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -17,46 +17,14 @@ const geistMono = Geist_Mono({
   subsets: ["latin"],
 });
 
-// export const metadata: Metadata = {
-//   title: "Create Next App",
-//   description: "Generated by create next app",
-// };
-
-function AuthProvider({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  const { checkAuth } = useAuthStore();
-
-  useEffect(() => {
-    checkAuth();
-  }, [checkAuth]);
-
-  return <>{children}</>;
-}
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  const path = usePathname();
-  const isAuthPage = path.startsWith("/auth");
-
+export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="en">
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
         suppressHydrationWarning={true}
       >
-        {isAuthPage ? (
-          <div className="bg-gray-50 min-h-screen">{children}</div>
-        ) : (
-          <AuthProvider>
-            <ProtectedRoute>{children}</ProtectedRoute>
-          </AuthProvider>
-        )}
+        <ClientLayout>{children}</ClientLayout>
       </body>
     </html>
   );
